+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ---
 ###5.9.4_WIP
 - Added Pushbullet notification if a new CS:GO update gets released
+- Added notification if the sent Tweet is protected
 
 ###5.9.3
 - Fixed selenium related features not working